Detached the file name from the canvas object
Previously, the file name would be attached to the canvas object, requiring the user to give the canvas a name. It also served the purpose of storing the file name when reading from a file. Now instead, the file name is returned along with the canvas but as a separate item inside a tuple. Not sure if getting the file name of the image being read is actually useful, but I'm going to keep the functionality in here just in case.